requirements.txt
Flask
flask-cors
pytesseract
Pillow
python-docx
openpyxl
reportlab
pypandoc
gunicorn